+++ api/integrations/gitlab/services/__init__.py
@@ -1,0 +1,25 @@
+from integrations.gitlab.services.tagging import (
+    apply_initial_tag,
+    apply_tag_for_event,
+    set_gitlab_tag,
+)
+from integrations.gitlab.services.url_parsing import (
+    parse_project_path,
+    parse_resource_iid,
+)
+from integrations.gitlab.services.webhooks import (
+    deregister_webhook_for_path,
+    ensure_webhook_registered,
+    has_live_resource_for_path,
+)
+
+__all__ = [
+    "apply_initial_tag",
+    "apply_tag_for_event",
+    "deregister_webhook_for_path",
+    "ensure_webhook_registered",
+    "has_live_resource_for_path",
+    "parse_project_path",
+    "parse_resource_iid",
+    "set_gitlab_tag",
+]
