+++ api/integrations/gitlab/services/webhooks.py
@@ -1,8 +1,5 @@
-import re
 import secrets
 import uuid
-from typing import Any
-from urllib.parse import urlparse
 
 import requests
 import structlog
@@ -13,41 +10,14 @@ from features.feature_external_resources.models import (
     GITLAB_RESOURCE_TYPES,
     FeatureExternalResource,
 )
-from features.models import Feature
 from integrations.gitlab.client import (
     create_project_hook,
     delete_project_hook,
 )
-from integrations.gitlab.constants import (
-    GITLAB_TAG_COLOR,
-    GITLAB_TAG_DESCRIPTION_BY_LABEL,
-    GITLAB_TAG_KIND_BY_LABEL,
-    GitLabTagLabel,
-)
-from integrations.gitlab.mappers import (
-    map_gitlab_resource_to_tag_label,
-    map_gitlab_webhook_payload_to_tag_label,
-    map_resource_url_to_filter_value,
-)
 from integrations.gitlab.models import GitLabConfiguration, GitLabWebhook
-from integrations.gitlab.types import GitLabWebhookPayload
-from projects.tags.models import Tag, TagType
+from integrations.gitlab.services.url_parsing import parse_project_path
 
 logger = structlog.get_logger("gitlab")
-
-_RESOURCE_PATH_PATTERN = re.compile(
-    r"^/(?P<path>.+?)/-/(?:issues|work_items|merge_requests)/\d+/?$"
-)
-
-
-def parse_project_path(resource_url: str | None) -> str | None:
-    """Return the GitLab project's URL-encodable path (``group/subgroup/project``)
-    from an issue or MR URL, or ``None`` if the URL isn't in the expected shape.
-    """
-    if not resource_url:
-        return None
-    match = _RESOURCE_PATH_PATTERN.match(urlparse(resource_url).path)
-    return match.group("path") if match else None
 
 
 def _get_webhook_url(webhook_uuid: uuid.UUID) -> str:
@@ -163,74 +133,3 @@ def deregister_webhook_for_path(
             gitlab__hook__id=webhook.gitlab_hook_id,
         )
         webhook.delete()
-
-
-def set_gitlab_tag(feature: Any, new_label: GitLabTagLabel) -> None:
-    """Apply a GitLab system tag to ``feature``, replacing any existing GitLab
-    tag of the same kind (Issue/MR) first. Issue tags and MR tags coexist
-    independently — an Issue event won't touch an MR tag or vice versa.
-    """
-    label_value = new_label.value
-
-    feature.tags.remove(
-        *feature.tags.filter(
-            type=TagType.GITLAB.value,
-            label__startswith=GITLAB_TAG_KIND_BY_LABEL[new_label],
-        )
-    )
-    tag, _ = Tag.objects.get_or_create(
-        label=label_value,
-        project=feature.project,
-        is_system_tag=True,
-        type=TagType.GITLAB.value,
-        defaults={
-            "color": GITLAB_TAG_COLOR,
-            "description": GITLAB_TAG_DESCRIPTION_BY_LABEL[new_label],
-        },
-    )
-    feature.tags.add(tag)
-
-
-def apply_initial_tag(resource: FeatureExternalResource) -> None:
-    """Tag `resource.feature` based on the linked GitLab issue/MR's state
-    at link time.
-    """
-    label = map_gitlab_resource_to_tag_label(resource)
-    if not label:
-        return
-    set_gitlab_tag(resource.feature, label)
-
-
-def apply_tag_for_event(
-    webhook: GitLabWebhook,
-    payload: GitLabWebhookPayload,
-) -> None:
-    if not (label := map_gitlab_webhook_payload_to_tag_label(payload)):
-        return
-
-    if not (
-        resource_url := (attrs := payload.get("object_attributes") or {}).get("url")
-    ):
-        return
-
-    if not (
-        feature := Feature.objects.filter(
-            project=webhook.gitlab_configuration.project,
-            external_resources__url__in=map_resource_url_to_filter_value(resource_url),
-            external_resources__type__in=GITLAB_RESOURCE_TYPES,
-        ).first()
-    ):
-        return
-
-    set_gitlab_tag(feature, label)
-    log = logger.bind(
-        organisation__id=webhook.gitlab_configuration.project.organisation_id,
-        project__id=webhook.gitlab_configuration.project_id,
-    )
-    log.info(
-        "feature.tagged",
-        feature__id=feature.id,
-        tag__label=label.value,
-        object_kind=payload.get("object_kind"),
-        action=attrs.get("action"),
-    )
