+++ api/tests/unit/integrations/gitlab/test_proxy_views.py
@@ -7,22 +7,11 @@ import requests
 from pytest_structlog import StructuredLogCapture
 from rest_framework import status
 
-from integrations.gitlab.models import GitLabConfiguration
-
 if TYPE_CHECKING:
     from pytest_mock import MockerFixture
     from rest_framework.test import APIClient
 
     from projects.models import Project
-
-
-@pytest.fixture()
-def gitlab_config(project: Project) -> GitLabConfiguration:
-    return GitLabConfiguration.objects.create(  # type: ignore[no-any-return]
-        project=project,
-        gitlab_instance_url="https://gitlab.example.com",
-        access_token="glpat-test-token",
-    )
 
 
 @pytest.mark.usefixtures("gitlab_config")
